refactor of MachineList as a component instead of a page
