+++ webapp/src/components/CustomNavbar/CustomNavbar.js
@@ -9,11 +9,6 @@ const CustomNavbar = () => {
             <Container>
                 <Navbar.Brand href="/dashboard" className="text-white" >NetCafé</Navbar.Brand>
                 <Navbar.Toggle aria-controls="custom-navbar" />
-                <Navbar.Collapse id="custom-navbar">
-                    <Nav>
-                        <Nav.Link href="/machines" className="text-white">Machines</Nav.Link>
-                    </Nav>
-                </Navbar.Collapse>
             </Container>
         </Navbar>
     )
